backend/main_fixed.py
# main.py - Fixed version
import functions_framework
from werkzeug.utils import secure_filename
import os
import json
import tempfile

# Import Google Cloud libraries with error handling
try:
    import google.auth
    from google.cloud import secretmanager
    from google.cloud import speech
    from google.cloud import texttospeech
    from google.oauth2.credentials import Credentials
    from google_auth_oauthlib.flow import Flow
    from googleapiclient.discovery import build
except ImportError as e:
    print(f"Import error: {e}")

# Configuration
SCOPES = ['https://www.googleapis.com/auth/gmail.readonly', 'https://www.googleapis.com/auth/calendar.readonly']
SECRET_NAME = 'oauth-client-secret'

@functions_framework.http
def multimodal_agent_orchestrator(request):
    """
    HTTP Cloud Function to orchestrate the AI agent's tasks.
    """
    if request.method != 'POST':
        return 'Method Not Allowed', 405

    try:
        # Basic response for now
        return {
            "status": "success", 
            "message": "Function deployed successfully with all imports!"
        }, 200
        
    except Exception as e:
        print(f"Error in orchestrator: {str(e)}")
        return f"Internal Server Error: {str(e)}", 500

def get_web_client_secrets():
    """Fetches the OAuth client secret from Google Secret Manager."""
    try:
        _, project_id = google.auth.default()
        client = secretmanager.SecretManagerServiceClient()
        name = f"projects/{project_id}/secrets/{SECRET_NAME}/versions/latest"
        response = client.access_secret_version(request={"name": name})
        return json.loads(response.payload.data.decode("UTF-8"))
    except Exception as e:
        print(f"Error getting client secrets: {e}")
        raise

def transcribe_audio(audio_bytes):
    """Transcribes audio bytes using the Speech-to-Text API."""
    try:
        client = speech.SpeechClient()
        audio = speech.RecognitionAudio(content=audio_bytes)
        config = speech.RecognitionConfig(
            encoding=speech.RecognitionConfig.AudioEncoding.AMR,
            sample_rate_hertz=8000,
            language_code="en-US",
        )
        response = client.recognize(config=config, audio=audio)
        
        if response.results and response.results[0].alternatives:
            return response.results[0].alternatives[0].transcript
        else:
            return ""
    except Exception as e:
        print(f"Error in transcription: {e}")
        return f"Transcription error: {str(e)}"

def synthesize_speech(text_to_speak, voice_choice="Standard"):
    """Synthesizes speech from text using the Text-to-Speech API."""
    try:
        client = texttospeech.TextToSpeechClient()
        synthesis_input = texttospeech.SynthesisInput(text=text_to_speak)

        if voice_choice == "WaveNet":
            voice_name = "en-US-Wavenet-D"
        else:
            voice_name = "en-US-Standard-C"

        voice = texttospeech.VoiceSelectionParams(
            language_code="en-US",
            name=voice_name
        )

        audio_config = texttospeech.AudioConfig(
            audio_encoding=texttospeech.AudioEncoding.MP3
        )

        response = client.synthesize_speech(
            input=synthesis_input, voice=voice, audio_config=audio_config
        )

        return response.audio_content
    except Exception as e:
        print(f"Error in speech synthesis: {e}")
        # Return a simple text response if TTS fails
        return text_to_speak.encode('utf-8')